kubam: fixed other test cases

+++ kubam/app/network/network.py
@@ -48,6 +48,11 @@ class Network(object):
         """
         Delete the network group from the config.
         """
+        if not isinstance(req, dict):
+            return {"error": "no parameters given to delete.  Please send the network name to delete"}, Const.HTTP_BAD_REQUEST
+        if not "name" in req:
+            return {"error": "please send the 'name' of the network to delete"}, Const.HTTP_BAD_REQUEST
+
         name = req['name']
         db = YamlDB()
         err, msg = db.delete_network_group(Const.KUBAM_CFG, name)
